requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# LightRAG Framework
lightrag-hku>=0.1.0  # Install with: pip install lightrag-hku

# OpenAI for embeddings and LLM
openai>=1.0.0
tiktoken>=0.5.0

# Data extraction
openpyxl>=3.1.0  # Excel files
pdfplumber>=0.9.0  # PDF files (future)

# Data anonymization
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.5.0

# Visualization
pyvis>=0.3.0  # Interactive graph visualization
networkx>=3.0  # Graph analysis
matplotlib>=3.7.0  # Charts

# Utilities
tqdm>=4.65.0  # Progress bars
colorama>=0.4.6  # Colored output
tabulate>=0.9.0  # Table formatting

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0

# Optional: Web search for advanced categorization
# tavily-python>=0.3.0